Alterando nomes de pastas, e resolvendo bugs do arduino, porém deixando prints para debug que devem ser removidos posteriormente
