--- arduino/FirmwareUpdater/Galileo Driver/linux-cdc-acm.inf
@@ -1,107 +1,0 @@
-; Windows USB CDC ACM Setup File
-
-; Based on INF template which was:
-;     Copyright (c) 2000 Microsoft Corporation
-; likely to be covered by the MLPL as found at:
-;    <http://msdn.microsoft.com/en-us/cc300389.aspx#MLPL>.
-; For use only on Windows operating systems.
-
-[Version]
-Signature="$Windows NT$"
-Class=Ports
-ClassGuid={4D36E978-E325-11CE-BFC1-08002BE10318}
-Provider=%Intel%
-DriverVer=09/15/2013,1.1.2600.0
-CatalogFile = Linux-cdc-acm.cat
-[Manufacturer]
-%Intel%=DeviceList, NTamd64
-
-[DestinationDirs]
-FakeModemCopyFileSection=12  
-DefaultDestDir=12
-
-
-;------------------------------------------------------------------------------
-;  Windows 2000/XP/Vista-32bit Sections
-;------------------------------------------------------------------------------
-
-[DriverInstall.nt]
-include=mdmcpq.inf
-;;CopyFiles=DriverCopyFiles.nt
- CopyFiles = FakeModemCopyFileSection 
-AddReg=DriverInstall.nt.AddReg
-
-
-
-[DriverInstall.nt.AddReg]
-HKR,,DevLoader,,*ntkern
-HKR,,NTMPDriver,,USBSER.sys
-HKR,,EnumPropPages32,,"MsPorts.dll,SerialPortPropPageProvider"
-
-[DriverInstall.nt.Services]
-AddService=usbser, 0x00000002, DriverService.nt
-
-[DriverService.nt]
-DisplayName=%SERVICE%
-ServiceType=1
-StartType=3
-ErrorControl=1
-ServiceBinary=%12%\USBSER.sys
-
-;------------------------------------------------------------------------------
-;  Vista-64bit Sections
-;------------------------------------------------------------------------------
-
-[DriverInstall.NTamd64]
-include=mdmcpq.inf
-;;CopyFiles=DriverCopyFiles.NTamd64
- CopyFiles = FakeModemCopyFileSection 
-AddReg=DriverInstall.NTamd64.AddReg
-
-
-
-[DriverInstall.NTamd64.AddReg]
-HKR,,DevLoader,,*ntkern
-HKR,,NTMPDriver,,USBSER.sys
-HKR,,EnumPropPages32,,"MsPorts.dll,SerialPortPropPageProvider"
-
-[DriverInstall.NTamd64.Services]
-AddService=usbser, 0x00000002, DriverService.NTamd64
-
-[DriverService.NTamd64]
-DisplayName=%SERVICE%
-ServiceType=1
-StartType=3
-ErrorControl=1
-ServiceBinary=%12%\USBSER.sys
-
-
-;------------------------------------------------------------------------------
-;  Vendor and Product ID Definitions
-;------------------------------------------------------------------------------
-; When developing your USB device, the VID and PID used in the PC side
-; application program and the firmware on the microcontroller must match.
-; Modify the below line to use your VID and PID.  Use the format as shown
-; below.
-; Note: One INF file can be used for multiple devices with different
-;       VID and PIDs.  For each supported device, append
-;       ",USB\VID_xxxx&PID_yyyy" to the end of the line.
-;------------------------------------------------------------------------------
-[SourceDisksFiles]
-[SourceDisksNames]
-[DeviceList]
-%DESCRIPTION%=DriverInstall, USB\VID_8086&PID_BABE
-
-[DeviceList.NTamd64]
-%DESCRIPTION%=DriverInstall, USB\VID_8086&PID_BABE
-
-
-;------------------------------------------------------------------------------
-;  String Definitions
-;------------------------------------------------------------------------------
-;Modify these strings to customize your device
-;------------------------------------------------------------------------------
-[Strings]
-Intel               = "Intel Corporation"
-DESCRIPTION         = "Galileo"
-SERVICE             = "USB RS-232 Emulation Driver"
