--- doc/instalacao.md
@@ -1,6 +1,0 @@
-Instalação do Ambiente de Desenvolvimento
-=========================================
-
-* Procedimento realizado nos seguintes SO's:
-
-  * `Windows 10;
