--- doc/tempo_aperto_rastro.txt
@@ -1,52 +1,0 @@
-Easy Slowest
-1,3924
-1,3903
-1,3868
-1,4097
-1,4022
-1,3987
-1,4193
-1,3981
-1,3996
-Média: 1,3997
-
-Easy Slower
-1,0364
-1,0737
-1,0668
-1,0406
-1,0665
-1,0696
-1,0341
-1,0472
-1,0269
-Média: 1,0513
-
-Easy Slow
-0,9277
-0,9026
-0,8353
-0,8645
-0,8753
-0,8904
-0,8644
-0,9187
-0,8367
-0,8936
-0,8654
-Média: 0,9675
-
-Easy Full Speed
-0,6636
-0,6692
-0,6602
-0,7083
-0,6803
-0,7008
-0,7037
-0,7259
-0,7230
-0,7128
-0,7141
-0,6694
-Média: 0,6940
