+++ docs/Cores.txt
@@ -1,0 +1,15 @@
+Notas
+Intervalo de valores RGB para detecção de notas pela cor
+            C. Máxima       C. Mínima
+Verde:      (0, 175, 0)     (255, 255, 255)
+Vermelho:   (175, 0, 0)     (255, 255, 255)
+Amarelo:    (175, 150, 0)   (255, 255, 255)
+Azul:       (0, 125, 175)   (255, 255, 255)
+Laranja:    (175, 95, 0)    (255, 255, 255)
+
+Cores centrais definidas para esferas de detecção de cor
+Verde:      (5, 255, 5)
+Vermelho:   (255, 97.5, 40)
+Amarelo:    (255, 230, 0)
+Azul:       (25, 217.5, 255)
+Laranja:    (255, 148, 10)
