+++ docs/instalacao.md
@@ -1,0 +1,6 @@
+Instalação do Ambiente de Desenvolvimento
+=========================================
+
+* Procedimento realizado nos seguintes SO's:
+
+  * `Windows 10;
