+++ docs/posicao_rastro
@@ -1,0 +1,72 @@
+% verde
+    imgO(293,238,:) = 255;
+    imgO(292,238,:) = 255;
+    imgO(291,239,:) = 255;
+    imgO(290,239,:) = 255;
+    imgO(289,239,:) = 255;
+    imgO(288,240,:) = 255;
+    imgO(287,240,:) = 255;
+    imgO(286,240,:) = 255;
+    imgO(285,241,:) = 255;
+    imgO(284,241,:) = 255;
+    imgO(283,241,:) = 255;
+    imgO(282,242,:) = 255;
+    imgO(281,243,:) = 255;
+%   vermelho
+    imgO(293,275,:) = 255;
+    imgO(292,275,:) = 255;
+    imgO(291,275,:) = 255;
+    imgO(290,276,:) = 255;
+    imgO(289,276,:) = 255;
+    imgO(288,276,:) = 255;
+    imgO(287,276,:) = 255;
+    imgO(286,276,:) = 255;
+    imgO(285,277,:) = 255;
+    imgO(284,277,:) = 255;
+    imgO(283,277,:) = 255;
+    imgO(282,277,:) = 255;
+    imgO(281,278,:) = 255;
+%   amarelo
+    imgO(293,312,:) = 255;
+    imgO(292,312,:) = 255;
+    imgO(291,312,:) = 255;
+    imgO(290,312,:) = 255;
+    imgO(289,312,:) = 255;
+    imgO(288,312,:) = 255;
+    imgO(287,312,:) = 255;
+    imgO(286,312,:) = 255;
+    imgO(285,312,:) = 255;
+    imgO(284,312,:) = 255;
+    imgO(283,312,:) = 255;
+    imgO(282,312,:) = 255;
+    imgO(281,312,:) = 255;
+
+%   azul
+    imgO(291,349,:) = 255;
+    imgO(290,349,:) = 255;
+    imgO(289,349,:) = 255;
+    imgO(288,349,:) = 255;
+    imgO(287,349,:) = 255;
+    imgO(286,349,:) = 255;
+    imgO(285,348,:) = 255;
+    imgO(284,348,:) = 255;
+    imgO(283,348,:) = 255;
+    imgO(282,348,:) = 255;
+    imgO(281,348,:) = 255;
+    imgO(280,348,:) = 255;
+    imgO(279,348,:) = 255;
+
+%   laranja
+    imgO(294,387,:) = 255;
+    imgO(293,387,:) = 255;
+    imgO(292,386,:) = 255;
+    imgO(291,386,:) = 255;
+    imgO(290,386,:) = 255;
+    imgO(289,385,:) = 255;
+    imgO(288,385,:) = 255;
+    imgO(287,385,:) = 255;
+    imgO(286,384,:) = 255;
+    imgO(285,384,:) = 255;
+    imgO(284,384,:) = 255;
+    imgO(283,383,:) = 255;
+    imgO(282,382,:) = 255;
